Set connection pool and update entities

+++ JavaEE_Rest_Service/TaskManager/src/main/java/com/aston/dao/api/UserDaoApi.java
@@ -6,7 +6,7 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface UserDaoApi {
-    int createUser(User user) throws  SQLException;
+    Long createUser(User user) throws  SQLException;
     User getUserById(int userId) throws  SQLException;
     User getUserByUsername(String username) throws  SQLException;
     List<User> getAllUsers() throws  SQLException;
