+++ JavaEE_Rest_Service/TaskManager/src/main/java/com/aston/util/dto/ProjectDto.java
@@ -1,12 +1,14 @@
 package com.aston.util.dto;
 
+import com.aston.entities.Project;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 public class ProjectDto {
-    private int id;
+    private Long id;
     private String name;
     private String description;
+
 }
